+++ compiler/compiler-plugin/src/main/kotlin/com/cmgapps/logtag/ir/LogTagIrGenerationExtension.kt
@@ -6,18 +6,21 @@
 
 package com.cmgapps.logtag.ir
 
+import dev.zacsweers.metro.compiler.compat.CompatContext
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
 import org.jetbrains.kotlin.ir.visitors.transformChildrenVoid
 
-class LogTagIrGenerationExtension : IrGenerationExtension {
+class LogTagIrGenerationExtension(
+    private val compatContext: CompatContext,
+) : IrGenerationExtension {
     override fun generate(
         moduleFragment: IrModuleFragment,
         pluginContext: IrPluginContext,
     ) {
         moduleFragment.transformChildrenVoid(
-            LogTagIrGenerator(pluginContext),
+            LogTagIrGenerator(pluginContext, compatContext),
         )
     }
 }